refactor: break out commands from entry file
